requirements.txt
Pillow
google-cloud-vision
pywin32
PySide6
pyclip